smarActMCS2MotorDriver.cpp: Major rework of report()
The "positioner name" was garbage:
strcpy(dst, src) it is.
Fix this, and while there, add a memset() and use strncpy() instead.
Add more variables to be retrieved and printed, range limit and thresholds.
To avoid too much code-duplication, add and use 2 helper functions to
simplify the code:
reportHelperInteger() and reportHelperDouble()