smarActMCS2MotorDriver.cpp: Support read only soft limits
The ESS version of the motor module, which includes the model 3 driver
and the motorRecord itself, support "read only softlimits".
If soft limits are set in the controller (and are valid, high > low),
they will be forwareded into the DLLM and DHLM fields of the motorRecord.

Additional logic inside the motorRecord will make sure that a user can not
override those values.

Support this logic and put it into #ifdef to enable this module to be
compiled against the upstream motor without this feature.