Network problems: Add writeReadOnErrorDisconnect() and us it
Detect network problems better when there is a query to the controller
and an answer is expected, but none is received.
It may happen, the the answer is stuck inside the network,
to be more exact inside the TCP layer, when there is a temporary network
problem.
To avoid that this answer is delivered later and treated as the
response to the next poll() and everyrhing is mixed up,
disconnect the socket (TCP connection).
asyn will the re-establish a new one without having old data floating
around.