smarActMCS2MotorDriver: Add handleStatusChange()
Partly revert the last commit:
The TCP/IP stack in the controller is only allowing one TCP connection
at a time. Trying to re-establish a new connection while the old one
is still not "fully closed" does not work.
This means that the "disconnect" from the previous commit may hang
the whold communication, if there is a short glitch in the network.
And the controller needs a power-cycle to get out of this situation.
The new solution will not close the TCP connection, but wait for
the controller to do so.
However, when polling one axis fails, all axes are put into the
disconnected state.
Where the initialPoll() will do the re-establishment.