smarActMCS2MotorDriver: Support openLoop
Add support for a probably unusual configuration:
The positioner has a sensor, but it should not be used.
The aim is to use an external readback to get the position
into motorRecord and use openloop movements in the MCS2.

Retries are handled by the motorRecord:
The fields RDBL, RTRY RMOD and URIP are used.
In the first phase of the project we can compare the external
readback with the sensor and see how good that works.
In the seconds phase the function of the sensor will be lost,
it is going to be damaged by radiation and stops working.

Add a new asyn parameter to be able to "connect" a record to it,
and use a local variable as a copy inside MCS2Axis::move().

This commit does not use the CNEN field from the motorRecord
for 2 reasons:
The default for the driver is still closed loop. The special case
open loop needs to be configured, when needed.
The CNEN field and specially the setClosedLoop() method in the
generic model 3 driver has been used to power on- and off the
motor. Espially the auto-power-on (and off) feature relies
on the fact that setClosedLoop() inside the driver enables the power
and nothing else.
The dilemma of CNEN used for different things has been realized
by the community, and my take away from this is to keep CNEN away
from closed loop.