Add VEL2CLF, Velocity to Frequency
The current implementation uses velocity (maxVelocity) in move() as the
directly as the frequency.
This has 2 disadvantages:
- The velocity from motorRecord (.VELO) is now longer reflected
  in EGU/second.
- Switching between closed- and open-loop gives a mess:
  A velocity of 1 mm/sec needs e.g a frequency of 300 Hz,
  as observed at one positioner.

To somewhat improve the situation, add an optional factor to calculate the
frequency from the velocity. This can be connected to an ao record.
When the new factor is 0.0, it is ignored, making the code backward
compatible with existing databases.