smarActMCS2MotorDriver: Add step frequency and count
Add 2 parameters (no records yet) that allow the driver to issue
a "step move" in open loop.
It turns out that all open loop step moves are using steps,
not picometers.

A better handling of open loop movements is wanted:
On the used hardware we can find that running +1000 moves the positioner
2.4 mm, and stepping backwards with -1000 moves -3.4 mm.

That will be addressed in the next commits.