smarActMCS2MotorDriver: Openloop: Add step size forward/reverse
Running the piezo in open-loop means that we use "(full) steps".
A step has a specific size, and the positioner even moves a slightly
different distance forward compared to reverse.
The the achieved step size may even be dependend on other things,
I don't know yet.
Anyway, make it possible to configure the step size into the driver,
so that the move() command achieves a good movement.
Example for records will be added later.