smarActMCS2MotorDriver: openloop: Cleanup vel2clf
Improve the calculation of the frequency when running open loop
and there is are step sizes defined.
Re-rorder the code to limit the scope of temporary variables.
Remove the parameter vel2clf because that is no longer needed.
Fix a bug: stepcnt is an integer to move the positioner a number of
steps, so create it as asynParamInt32, as it should be.