smarActMCS2MotorDriver: Replace MCS2Controller::writeInt32
Make the driver more in line which what is intended to be overridde.
Remove MCS2Controller::writeInt32() and add MCS2Axis::setIntegerParam()

The writeInt32() method from the controller could/should be left as is,
instead of duplicating the code to find out the function from
the asynUser, call the corresponding axis function and do
the callbacks.

The intention is to override setIntegerParam() in the axis class
instead, which is done in this commit.

Note that calling things like
setIntegerParam(pC_->motorStatusDone_, done);
from inside the axis can be simplified:
Instead of using setIntegerParam() from MCS2Axis with all the
if "function == this" or "function == that" we can use the
base class:
asynMotorAxis::setIntegerParam(pC_->motorStatusDone_, done);

Note: The same optimization can be done for setDoubleParam(),
in a different commit.